minor cast fixes to prevent 64bit warnings

+++ generic/tkPack.c
@@ -10,7 +10,7 @@
  * See the file "license.terms" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * RCS: @(#) $Id: tkPack.c,v 1.10 2001/09/17 23:16:26 dgp Exp $
+ * RCS: @(#) $Id: tkPack.c,v 1.11 2001/09/21 21:34:10 hobbs Exp $
  */
 
 #include "tkPort.h"
@@ -1214,19 +1214,19 @@ PackAfter(interp, prevPtr, masterPtr, objc, objv)
 	    c = curOpt[0];
 
 	    if ((c == 't')
-		    && (strncmp(curOpt, "top", length)) == 0) {
+		    && (strncmp(curOpt, "top", (size_t) length)) == 0) {
 		packPtr->side = TOP;
 	    } else if ((c == 'b')
-		    && (strncmp(curOpt, "bottom", length)) == 0) {
+		    && (strncmp(curOpt, "bottom", (size_t) length)) == 0) {
 		packPtr->side = BOTTOM;
 	    } else if ((c == 'l')
-		    && (strncmp(curOpt, "left", length)) == 0) {
+		    && (strncmp(curOpt, "left", (size_t) length)) == 0) {
 		packPtr->side = LEFT;
 	    } else if ((c == 'r')
-		    && (strncmp(curOpt, "right", length)) == 0) {
+		    && (strncmp(curOpt, "right", (size_t) length)) == 0) {
 		packPtr->side = RIGHT;
 	    } else if ((c == 'e')
-		    && (strncmp(curOpt, "expand", length)) == 0) {
+		    && (strncmp(curOpt, "expand", (size_t) length)) == 0) {
 		packPtr->flags |= EXPAND;
 	    } else if ((c == 'f')
 		    && (strcmp(curOpt, "fill")) == 0) {
@@ -1264,7 +1264,7 @@ PackAfter(interp, prevPtr, masterPtr, objc, objv)
 		packPtr->iPadY = 0;
 		index++;
 	    } else if ((c == 'f') && (length > 1)
-		    && (strncmp(curOpt, "frame", length) == 0)) {
+		    && (strncmp(curOpt, "frame", (size_t) length) == 0)) {
 		if (optionCount < (index+2)) {
 		    Tcl_AppendResult(interp, "wrong # args: \"frame\" ",
 			    "option must be followed by anchor point",
